+++ force_wfmanager/gui/tests/test_click_run.py
@@ -12,6 +12,15 @@ def mock_run_constructor(*args, **kwargs):
     mock_wf_run.main = lambda: None
 
 
+class MockWfManager(object):
+
+    def __init__(self, *args, **kwargs):
+        pass
+
+    def run(self):
+        pass
+
+
 class TestClickRun(unittest.TestCase):
 
     def test_click_cli_version(self):
@@ -27,3 +36,10 @@ class TestClickRun(unittest.TestCase):
             force_wfmanager.gui.run.force_wfmanager()
 
             self.assertTrue(mock_run.force_wfmanager.called)
+
+    def test_run_with_debug(self):
+        with mock.patch('force_wfmanager.gui.run.WfManager') as mock_wf:
+            mock_wf.return_value = MockWfManager()
+            force_wfmanager.gui.run.main(debug=True)
+            self.log = force_wfmanager.gui.run.logging.getLogger(__name__)
+            self.assertEqual(self.log.getEffectiveLevel(), 10)